Write initial experimental module descriptors
This provides module-info.java files for the core and io
modules. These absolutely should *NOT* make it into any release,
because the io descriptor has a "requires" clause that mentions
the automatically-generated "antlr4.runtime" name as opposed to the
correct "org.antlr.antlr4.runtime" name that must be used if and when
the Antlr4 Automatic-Module-Name is published. The old name
is used purely to test whether or not the code works when modularized.

See https://github.com/antlr/antlr4/pull/2223

Affects #484